Add delete function test
