+++ src/__mocks__/taskUtils.js
@@ -4,10 +4,6 @@ const mock = jest.fn().mockImplementation(() => ({
   getTasks() {
     return tasks;
   },
-  deleteTask(index) {
-    tasks.filter((todo, i) => i !== index);
-    tasks.splice(index, 1);
-  },
   editTaskDescription(task) {
     const index = task.index - 1;
     tasks[index].description = task.description;
